requirements.txt
streamlit>=1.33.0
requests>=2.31.0
httpx>=0.27.0
feedparser>=6.0.11
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-mock>=3.14.0
